Restructured into class. Added header and footer for index.php and setup.php Added "Searched for" section for users to know what they searched
