--- example.php
@@ -1,14 +1,0 @@
-<?php
-    require('dbSearcher/core.php'); // Load DBS-Core
-?>
-
-<html>
-    <head>
-        <title>Database Searcher Example</title>
-        <meta charset="UTF-8">
-    </head>
-    <body>
-        <!-- Generate DBS markup view -->
-        <?php _DBSearcher();?>
-    </body>
-</html>
