+++ footer.php
@@ -1,0 +1,1 @@
+</body></html>
