+++ header.php
@@ -1,0 +1,13 @@
+<?php
+    require('dbSearcher/dbSearcher.php'); // Load DBS-Core
+    
+    $dbSearcher = new dbSearcher();
+?>
+
+<html>
+    <head>
+        <title>Database Searcher</title>
+        <meta charset="UTF-8">
+        <link rel="stylesheet" href="css/styles.css" />
+    </head>
+    <body>
